refactor(plugins): rename node-only plugin loader to plugin-loader.node.ts
Make the runtime home of the plugin discovery/loading module explicit.

`src/plugins/index.ts` touches Node-only APIs directly: `node:fs`,
`node:path`, `process.env['HOME']`, `process.env['INSOMNIA_DATA_PATH']`,
`require`, and `electron`. With contextIsolation now enabled on the main
BrowserWindow (#10111), these would silently break if reached from the main
renderer. They don't break today only because all callers run in Node-capable
contexts (Electron main, the node network/utility runtime, the inso CLI, and
the hidden plugin window which keeps `nodeIntegration: true`); the main
renderer talks to these functions over the plugin IPC bridge
(`~/ui/plugins/renderer-bridge`).

The neutral `plugins/index.ts` name and a stale vite comment calling it a
"renderer file" obscured that. This renames it to `plugin-loader.node.ts`,
matching the repo's existing `.node.ts` convention for node-only modules
(e.g. `send-request.node.ts`, `write-proto-file.node.ts`), and adds a header
comment documenting the constraint. No behavior change.

- Rename src/plugins/index.ts -> src/plugins/plugin-loader.node.ts
- Rename its test index.test.ts -> plugin-loader.node.test.ts
- Update importers: invoke-method.ts, main/templating-worker-database.ts,
  runtimes/network/network-adapter.node.ts, templating/index.ts, and tests
- Fix stale "renderer file / contextIsolation currently false" comment in
  vite.config.ts
